Fix: The lack of the class path for URLClassLoader

+++ nesp-gradle-plugin-javafx/src/main/java/com/nesp/gradle/plugin/javafx/JavaFxPlugin.java
@@ -21,7 +21,6 @@ import com.nesp.gradle.plugin.javafx.fxml.GenerateBaseControllerClassFileTask;
 import com.nesp.gradle.plugin.javafx.resource.GenerateRClassFileTask;
 import com.nesp.gradle.plugin.javafx.resource.ResourceConfig;
 import com.nesp.gradle.plugin.javafx.utils.ProjectUtils;
-import org.apache.tools.ant.util.FileUtils;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
@@ -30,6 +29,7 @@ import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.tasks.compile.JavaCompile;
 
+import javax.annotation.Nonnull;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
@@ -38,12 +38,33 @@ import java.util.Set;
 
 public class JavaFxPlugin implements Plugin<Project> {
 
+    public static final boolean DEBUG = false;
+
     public static final String NESP_JAVA_FX_PLUGIN_EXTENSION_NAME = "nespJfx";
     public static final String GENERATE_BASE_CONTROLLER_FILE_TASK_NAME = "generateJavaFxBaseControllerFile";
     public static final String GENERATE_R_FILE_TASK_NAME = "generateJavaFxRFile";
 
+    public static void printDebugLog(final String tag, final String msg) {
+        printLog(true, tag, msg);
+    }
+
+    public static void printLog(final String tag, final String msg) {
+        printLog(false, tag, msg);
+    }
+
+    private static void printLog(final boolean isDebug, final String tag, final String msg) {
+        if (isDebug && !DEBUG) return;
+        final StringBuilder log = new StringBuilder(NESP_JAVA_FX_PLUGIN_EXTENSION_NAME).append(": ");
+        if (tag == null || tag.isEmpty()) {
+            log.append(msg);
+        } else {
+            log.append(tag).append(": ").append(msg);
+        }
+        System.out.println(log);
+    }
+
     @Override
-    public void apply(Project project) {
+    public void apply(@Nonnull Project project) {
         configureSourceSetGenerateFileOutputs(project);
         configureNespJavaFxPluginExtension(project);
         configureGenerateBaseControllerFileTask(project);
@@ -55,13 +76,14 @@ public class JavaFxPlugin implements Plugin<Project> {
      *
      * @param project project
      */
+    @SuppressWarnings("all")
     private void configureSourceSetGenerateFileOutputs(Project project) {
         // Do not use lambdas.
         // Using Java lambdas is not supported as task inputs.
         // Please refer to https://docs.gradle.org/7.2/userguide/validation_problems.html#implementation_unknown
         project.afterEvaluate(new Action<Project>() {
             @Override
-            public void execute(final Project project1) {
+            public void execute(@Nonnull final Project project1) {
                 if (project1.getPlugins().hasPlugin(JavaPlugin.class)) {
                     final File desFile = new File(ProjectUtils.getGenerateSourcePath(project));
 
@@ -83,7 +105,7 @@ public class JavaFxPlugin implements Plugin<Project> {
 
                         mainSrcDirs.add(desFile);
 
-                        System.out.println("add src dir: " + desFile.getAbsolutePath());
+//                        printDebugLog("Add src dir", desFile.getAbsolutePath());
 
                         javaPluginExtension
                                 .getSourceSets()
@@ -96,18 +118,22 @@ public class JavaFxPlugin implements Plugin<Project> {
         });
     }
 
+    @SuppressWarnings("unused")
     private void deleteDirectory(final File desFile) {
         final File[] files = desFile.listFiles();
-        if (files != null) {
+        if (files != null && files.length > 0) {
             for (final File file : files) {
                 if (file.isFile()) {
-                    final boolean _d = file.delete();
+                    if (!file.delete()) {
+                        printLog("deleteDirectory", "file " + file + " delete failed");
+                    }
                 } else {
                     deleteDirectory(file);
                 }
             }
-            if (desFile.listFiles() == null || desFile.listFiles().length == 0) {
-                desFile.delete();
+        } else {
+            if (!desFile.delete()) {
+                printLog("deleteDirectory", "file " + desFile + " delete failed");
             }
         }
     }
@@ -116,13 +142,14 @@ public class JavaFxPlugin implements Plugin<Project> {
         project.getExtensions().add(NESP_JAVA_FX_PLUGIN_EXTENSION_NAME, NespJavaFxPluginExtension.class);
     }
 
+    @SuppressWarnings("all")
     private void configureGenerateBaseControllerFileTask(Project project) {
         // Do not use lambdas.
         // Using Java lambdas is not supported as task inputs.
         // Please refer to https://docs.gradle.org/7.2/userguide/validation_problems.html#implementation_unknown
         project.afterEvaluate(new Action<Project>() {
             @Override
-            public void execute(final Project project1) {
+            public void execute(@Nonnull final Project project1) {
                 NespJavaFxPluginExtension nespJfx =
                         (NespJavaFxPluginExtension) project1.getExtensions()
                                 .findByName(NESP_JAVA_FX_PLUGIN_EXTENSION_NAME);
@@ -145,7 +172,7 @@ public class JavaFxPlugin implements Plugin<Project> {
                     JavaCompile compileJavaTask = (JavaCompile) compileJavaTasks.toArray(new Object[0])[0];
                     compileJavaTask.doFirst(new Action<Task>() {
                         @Override
-                        public void execute(final Task task1) {
+                        public void execute(@Nonnull final Task task1) {
                             if (baseControllerEnable) {
                                 generateBaseControllerClassFileTask.run();
                             }
@@ -156,13 +183,14 @@ public class JavaFxPlugin implements Plugin<Project> {
         });
     }
 
+    @SuppressWarnings("all")
     private void configureGenerateRFileTask(Project project) {
         // Do not use lambdas.
         // Using Java lambdas is not supported as task inputs.
         // Please refer to https://docs.gradle.org/7.2/userguide/validation_problems.html#implementation_unknown
         project.afterEvaluate(new Action<Project>() {
             @Override
-            public void execute(final Project project1) {
+            public void execute(@Nonnull final Project project1) {
                 NespJavaFxPluginExtension nespJfx =
                         (NespJavaFxPluginExtension) project1.getExtensions()
                                 .findByName(NESP_JAVA_FX_PLUGIN_EXTENSION_NAME);
@@ -178,7 +206,7 @@ public class JavaFxPlugin implements Plugin<Project> {
                     JavaCompile compileJavaTask = (JavaCompile) compileJavaTasks.toArray(new Object[0])[0];
                     compileJavaTask.doFirst(new Action<Task>() {
                         @Override
-                        public void execute(final Task task1) {
+                        public void execute(@Nonnull final Task task1) {
                             generateRClassFileTask.run();
                         }
                     });
